Encapsulating Gender enumeration.
This commit introduces a new modularity feature to the code. It includes a enumeration for codificating available genders (Male and Female), and a new service that include functions for automatically generating radio input in order to select a gender during sign up or editions. It also includes the subsequent modifications for User entity and for its sample data loader.