Encapsulating RefereeRange enumeration.
This commit introduces a new modularity feature to the code. It includes a enumeration for codificating available ranges, and a new service that include functions for automatically generating select input in order to choose a range during sign up or editions. It also includes the subsequent modifications for User entity and for its sample data loader.