Encapsulating roles of application in a enumeration.
This commit introduces a new modularity feature to the code. It includes a enumeration for codificating available roles (competitor, referee and administrator). It has been designed thinking about security aspects (like banning access to specific pages or including different elements in dynamic views). It also includes the subsequent modifications for User entity (which now allows to assign multiple roles for a user) and for its sample data loader.